refactor: parseNews.ts 추가
